refactor: streamline custom blocks build/lint process
- move example blocks into subfolders of src directory
- streamline build process for example blocks
- support view scripts in example blocks
- move wp-scripts to top-level package.json
- delete unnecessary dotfiles
- update instructions for using/writing custom blocks
- use same linters for plugins and other JS/SCSS
- remove formatting script of limited use
