+++ .github/workflows/validation.yml
@@ -50,9 +50,6 @@ jobs:
       - name: Install npm dependencies
         run: npm ci
 
-      - name: Install custom plugin npm dependencies
-        run: npm run plugins:install
-
       - name: Lint
         run: npm run lint
         env:
