+++ package.json
@@ -19,23 +19,19 @@
     "postsass": "postcss theme/css/base.css --use autoprefixer -r",
     "js:watch": "NODE_ENV=development node esbuild.mjs",
     "js:prod": "node esbuild.mjs",
-    "plugins:install": "cd src/plugins/example-blocks && npm ci",
-    "plugins:dev": "cd src/plugins/example-blocks && npm start",
-    "plugins:build": "cd src/plugins/example-blocks && npm run build",
-    "plugins:format": "cd src/plugins/example-blocks && npm run format",
+    "plugins:dev": "wp-scripts start --webpack-src-dir=src/plugins/example-blocks/src --output-path=src/plugins/example-blocks/build",
+    "plugins:build": "wp-scripts build --webpack-src-dir=src/plugins/example-blocks/src --output-path=src/plugins/example-blocks/build",
     "watch": "run-p watch:*",
     "watch:php": "chokidar \"src/php/**/*\" -c \"npm run copy:php\"",
     "watch:scss": "chokidar \"src/scss/**/*\" -c \"npm run sass\"",
     "test": "vitest --run",
     "test:watch": "vitest",
     "test:ci": "vitest --run --coverage",
-    "lint:plugins:js": "cd src/plugins/example-blocks && npm run lint:js",
-    "lint:plugins:css": "cd src/plugins/example-blocks && npm run lint:css",
-    "lint:js": "eslint src/js",
+    "lint:js": "eslint src",
     "lint:php": "./scripts/run.sh vendor/bin/phpcs --standard=wp-configs/phpcs-rules-standard.xml --extensions=php src/php -s -n",
-    "lint:scss": "stylelint src/scss/**/*.scss src/scss/*.scss --color --formatter verbose",
+    "lint:scss": "stylelint \"src/**/*.scss\" --color --formatter verbose",
     "lint:twig": "./scripts/run.sh vendor/bin/twigcs src/php/views",
-    "lint": "run-s lint:* lint:plugins:*",
+    "lint": "run-s lint:*",
     "php:run": "./scripts/run.sh",
     "export-db": "./scripts/export-db.sh",
     "backup-db": "BACKUP=true ./scripts/export-db.sh",
@@ -58,6 +54,7 @@
     "@testing-library/dom": "^9.3.3",
     "@testing-library/jest-dom": "^6.1.4",
     "@vitest/coverage-istanbul": "^0.34.6",
+    "@wordpress/scripts": "^26.15.0",
     "autoprefixer": "^10.4.16",
     "check-node-version": "^4.2.1",
     "chokidar-cli": "^3.0.0",
