--- src/plugins/example-blocks/.prettierrc
@@ -1,14 +1,0 @@
-{
-	"printWidth": 110,
-	"endOfLine": "lf",
-	"trailingComma": "es5",
-	"tabWidth": 2,
-	"useTabs": true,
-	"semi": true,
-	"singleQuote": true,
-	"quoteProps": "as-needed",
-	"jsxSingleQuote": false,
-	"bracketSpacing": true,
-	"bracketSameLine": false,
-	"arrowParens": "always"
-}
