--- src/plugins/example-blocks/package.json
@@ -1,20 +1,0 @@
-{
-	"name": "example-blocks",
-	"version": "0.1.0",
-	"description": "Collection of example blocks.",
-	"author": "The WordPress Contributors",
-	"license": "GPL-2.0-or-later",
-	"main": "build/index.js",
-	"scripts": {
-		"build": "./scripts/build-blocks.sh",
-		"format": "wp-scripts format",
-		"lint:css": "wp-scripts lint-style",
-		"lint:js": "wp-scripts lint-js",
-		"packages-update": "wp-scripts packages-update",
-		"plugin-zip": "wp-scripts plugin-zip",
-		"start": "./scripts/start-blocks.sh"
-	},
-	"devDependencies": {
-		"@wordpress/scripts": "^26.13.0"
-	}
-}
