+++ src/plugins/example-blocks/src/hello-world/editor.scss
@@ -5,6 +5,5 @@
  */
 
 .wp-block-example-blocks-hello-world {
-
 	/* insert custom styles here */
 }
