+++ src/plugins/example-blocks/src/hello-world/index.js
@@ -19,7 +19,7 @@ import './style.scss';
  */
 import Edit from './edit';
 import save from './save';
-import metadata from '../block.json';
+import metadata from './block.json';
 
 /**
  * Every block starts by registering a new block type definition.
