+++ src/plugins/example-blocks/src/hello-world/view.js
@@ -1,0 +1,1 @@
+console.log('Hello from the Hello World block!');
