+++ src/plugins/example-blocks/src/reverse-string/block.json
@@ -11,7 +11,8 @@
 		"html": false
 	},
 	"textdomain": "example-blocks",
-	"editorScript": "file:./build/index.js",
-	"editorStyle": "file:./build/index.css",
-	"style": "file:./build/style-index.css"
+	"editorScript": "file:index.js",
+	"editorStyle": "file:index.css",
+	"style": "file:style-index.css",
+	"viewScript": "file:view.js"
 }
