+++ src/plugins/example-blocks/src/reverse-string/edit.js
