+++ src/plugins/example-blocks/src/reverse-string/editor.scss
