+++ src/plugins/example-blocks/src/reverse-string/style.scss
@@ -6,6 +6,5 @@
  */
 
 .wp-block-example-blocks-reverse-string {
-
 	/* insert custom styles here */
 }
