+++ src/plugins/example-blocks/src/reverse-string/view.js
@@ -1,0 +1,1 @@
+console.log('Hello from the Reverse String block!');
